requirements.txt
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.58.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
isort==6.0.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaleido==1.0.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-anthropic==0.3.15
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
logistro==1.1.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
minio==7.2.15
mypy_extensions==1.1.0
narwhals==1.42.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.6
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
pip-chill==1.0.3
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
tenacity==9.1.2
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
zstandard==0.23.0
